chore: Pull out normalize logic from cargo copy
